stt-sample/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "openai-whisper>=20250625",
    "pydantic>=2.11.7",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
    "uvicorn>=0.35.0",
]
